requirements.txt
PyYaml~=6.0
setuptools==67.7.2
beautifulsoup4~=4.12.2